move btc2sat and sat2btc out of core/utils
