+++ src/core/specter-wallet.ts
@@ -6,9 +6,8 @@ import { getActiveOnchainLnd, lndsBalances } from "@services/lnd/utils"
 import { ledger } from "@services/mongodb"
 
 import { UserWallet } from "./user-wallet"
-import { btc2sat, sat2btc } from "./utils"
 import { BTC_NETWORK, ONCHAIN_SCAN_DEPTH_OUTGOING } from "@config/app"
-import { toSats } from "@domain/bitcoin"
+import { btc2sat, sat2btc, toSats } from "@domain/bitcoin"
 import { OnChainService } from "@services/lnd/onchain-service"
 import { TxDecoder } from "@domain/bitcoin/onchain"
