+++ test/integration/02-user-wallet/02-receive-onchain.spec.ts
@@ -1,6 +1,6 @@
 import { once } from "events"
 import { getOnChainAddressCreateAttemptLimits, getUserLimits } from "@config/app"
-import { btc2sat, sat2btc, sleep } from "@core/utils"
+import { sleep } from "@core/utils"
 import { getTitle } from "@services/notifications/payment"
 import { onchainTransactionEventHandler } from "@servers/trigger"
 import {
@@ -26,6 +26,7 @@ import { baseLogger } from "@services/logger"
 import { getFunderWalletId } from "@services/ledger/accounts"
 import { WalletFactory } from "@core/wallet-factory"
 import { User } from "@services/mongoose/schema"
+import { btc2sat, sat2btc } from "@domain/bitcoin"
 
 jest.mock("@app/prices/get-current-price", () => require("test/mocks/get-current-price"))
